Develop

+++ src/main/java/com/shepard1992/gmail/twoya_formochka/repository/api/ClientRepository.java
@@ -5,7 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
+import java.awt.*;
+
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Long>, JpaSpecificationExecutor<Client> {
+public interface ClientRepository extends JpaRepository<Client, Integer>, JpaSpecificationExecutor<Client> {
 
 }
